migrated user-report template
