+++ email-templates/user-report.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="{{lang}}">
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>User Activity Report</title>
+    <title>{{title}}</title>
     <style>
     @page {
       size: A4;
@@ -179,6 +179,7 @@
         page-break-inside: avoid;
       }
     }
+    {{customCSS}}
   </style>
   </head>
   <body>
